requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-bootstrap-modal-forms==2.2.0
django-ckeditor==6.2.0
django-js-asset==1.2.2
django-mptt==0.13.4
djangorestframework==3.13.1
Pillow==8.4.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
